feat:Update contact pages to be dynamic
