--- graphql/queries/contact-canada-pension-plan.graphql
@@ -1,133 +1,0 @@
-{
-    schPagev1ByPath(
-    _path: "/content/dam/decd-endc/content-fragments/sch/pages/contact-us-pages/cpp-contacts"
-  ) {
-    item {
-      _path
-      scId
-      scTitleEn
-      scTitleFr
-      scPageNameEn
-      scPageNameFr
-      scShortTitleEn
-      scShortTitleFr
-      scBreadcrumbParentPages {
-          _path
-          scId
-          scPageNameEn
-          scPageNameFr
-          scTitleEn
-          scTitleFr
-          scShortTitleFr
-          scShortTitleEn
-          scOwner
-          scDateModifiedOverwrite
-          scAudience
-      }
-      scContentType
-      scAudience
-      scOwner
-      scFragments {
-          ... on SCHListv1Model {
-              _path
-              scId
-              scTitleEn
-              scTitleFr
-              scItems {
-                  ... on SCHContactMethodv1Model {
-                      _path
-                      scId
-                      scTitleEn
-                      scTitleFr
-                      schContactMethodLayout
-                      schIntroEn {
-                          markdown
-                      }
-                      schIntroFr {
-                          markdown
-                      }
-                      schHoursDetailsFr {
-                          markdown
-                      }
-                      schLocationLabelEn {
-                          markdown
-                      }
-                      schDetails 
-                          {
-                              ... on SCHListv1Model {
-                                  _path
-                                  scBackgroundColour
-                                  scTitleEn
-                                  scTitleFr
-                                   scId
-                                   scItems {
-                                     ... on SCHAddressv1Model {
-                                              _path
-                                              scId
-                                              scRecipientEn
-                                              scRecipientFr
-                                              scPostalBoxEn
-                                              scPostalBoxFr
-                                              scPostalStationEn
-                                              scPostalStationFr
-                                              scCityEn
-                                              scCityFr
-                                              scProvTerrAbbr
-                                              scPostalCode
-                                              scCountry
-                                              scProvTerrAbbrEnum
-                                              scCountryEn
-                                              scCountryFr
-                                          }
-                                      ... on SCHContentv1Model {
-                                      _path
-                                      scId
-                                      scContentEn {
-                                          markdown
-                                      }
-                                      scContentFr {
-                                          markdown
-                                      }
-                                      scIconCSS
-                                      scFragments {
-                                          ... on SCHTaskv1Model {
-                                              scId
-                                              scLinkTextEn
-                                              scLinkTextFr
-                                              scDestinationURLEn
-                                              scDestinationURLFr
-                                              scButtonType
-                                          }
-                                          ... on SCHAddressv1Model {
-                                              _path
-                                              scId
-                                              scRecipientEn
-                                              scRecipientFr
-                                              scProgramEn
-                                              scProgramFr
-                                              scPostalBoxEn
-                                              scPostalBoxFr
-                                              scPostalStationEn
-                                              scPostalStationFr
-                                              scCityEn
-                                              scCityFr
-                                              scProvTerrAbbr
-                                              scPostalCode
-                                              scCountry
-                                              scProvTerrAbbrEnum
-                                              scCountryEn
-                                              scCountryFr
-                                          }
-                                      }
-                                  }
-                                  }
-                              }
-                          }
-                  }
-                }
-
-          }
-      }
-    }
-  }
-}
