--- pages/contact-us.js
@@ -1,168 +1,0 @@
-import PropTypes from 'prop-types'
-import { Heading, Link } from '@dts-stn/service-canada-design-system'
-import en from '../locales/en'
-import fr from '../locales/fr'
-import { getContactUsContent } from '../graphql/mappers/contact-us'
-import { getBetaBannerContent } from '../graphql/mappers/beta-banner-opt-out'
-import { getBetaPopupExitContent } from '../graphql/mappers/beta-popup-exit'
-import logger from '../lib/logger'
-import { useEffect, useCallback, useMemo } from 'react'
-import throttle from 'lodash.throttle'
-import NextLink from 'next/link'
-
-export default function ContactLanding(props) {
-  const t = props.locale === 'en' ? en : fr
-
-  //Event listener for click events that revalidates MSCA session, throttled using lodash to only trigger every 15 seconds
-  const onClickEvent = useCallback(() => fetch('/api/refresh-msca'), [])
-  const throttledOnClickEvent = useMemo(
-    () => throttle(onClickEvent, 15000, { trailing: false }),
-    [onClickEvent]
-  )
-
-  useEffect(() => {
-    window.addEventListener('click', throttledOnClickEvent)
-    //Remove event on unmount to prevent a memory leak with the cleanup
-    return () => {
-      window.removeEventListener('click', throttledOnClickEvent)
-    }
-  }, [throttledOnClickEvent])
-
-  return (
-    <div id="contactContent" data-testid="contactContent-test">
-      <Heading id="my-dashboard-heading" title={props.content.heading} />
-      <p className="mt-3 mb-8 text-xl font-body">{props.content.subHeading}</p>
-      <ul className="list-disc" data-cy="contact-task-list">
-        {props.content.links.map((link) => {
-          return (
-            <li className="mb-6 ml-5" key={link.linkId}>
-              <Link
-                id={link.linkId}
-                dataTestId={link.linkId}
-                text={link.linkTitle}
-                href={`/${props.content.pageName}/${link.linkDestination
-                  .split('/')
-                  .pop()}`}
-                component={NextLink}
-              />
-              <p className="text-xl font-body">{link.linkDescription}</p>
-            </li>
-          )
-        })}
-      </ul>
-    </div>
-  )
-}
-
-export async function getServerSideProps({ res, locale }) {
-  const content = await getContactUsContent().catch((error) => {
-    logger.error(error)
-    //res.statusCode = 500
-    throw error
-  })
-  const bannerContent = await getBetaBannerContent().catch((error) => {
-    logger.error(error)
-    // res.statusCode = 500
-    throw error
-  })
-  const popupContent = await getBetaPopupExitContent().catch((error) => {
-    logger.error(error)
-    // res.statusCode = 500
-    throw error
-  })
-
-  /* 
-   * Uncomment this block to make Banner Popup Content display "Page Not Available"
-   * Comment "getBetaPopupExitContent()" block of code above.
-  
-    const popupContent = await getBetaPopupNotAvailableContent().catch((error) => {
-      logger.error(error)
-      // res.statusCode = 500
-      throw error
-    })
-  */
-
-  /* istanbul ignore next */
-  const langToggleLink =
-    locale === 'en' ? '/fr/contactez-nous' : '/en/contact-us'
-
-  const t = locale === 'en' ? en : fr
-
-  const breadCrumbItems =
-    locale === 'en'
-      ? content.en.breadcrumb?.map(({ link, text }) => {
-          return { text, link: '/' + locale + '/' + link }
-        })
-      : content.fr.breadcrumb?.map(({ link, text }) => {
-          return { text, link: '/' + locale + '/' + link }
-        })
-
-  /* Place-holder Meta Data Props */
-  const meta = {
-    data_en: {
-      title: 'Contact - My Service Canada Account - Contact',
-      desc: 'English',
-      author: 'Service Canada',
-      keywords: '',
-      service: 'ESDC-EDSC_MSCA-MSDC',
-      creator: 'Employment and Social Development Canada',
-      accessRights: '1',
-    },
-    data_fr: {
-      title: 'Contactez-nous - Mon dossier Service Canada',
-      desc: 'Français',
-      author: 'Service Canada',
-      keywords: '',
-      service: 'ESDC-EDSC_MSCA-MSDC',
-      creator: 'Emploi et Développement social Canada',
-      accessRights: '1',
-    },
-  }
-
-  return {
-    props: {
-      locale,
-      langToggleLink,
-      content: locale === 'en' ? content.en : content.fr,
-      meta,
-      breadCrumbItems,
-      bannerContent: locale === 'en' ? bannerContent.en : bannerContent.fr,
-      popupContent: locale === 'en' ? popupContent.en : popupContent.fr,
-      aaPrefix: `ESDC-EDSC:${content.en.heading}`,
-    },
-  }
-}
-
-ContactLanding.propTypes = {
-  /**
-   * current locale in the address
-   */
-  locale: PropTypes.string,
-
-  /*
-   * Language link toggle text
-   */
-  langToggleLink: PropTypes.string,
-
-  /*
-   * Content Tags
-   */
-
-  content: PropTypes.object,
-
-  /*
-   * Meta Tags
-   */
-
-  meta: PropTypes.object,
-
-  /*
-   * BreadCrumb Items
-   */
-  breadCrumbItems: PropTypes.arrayOf(
-    PropTypes.shape({
-      text: PropTypes.string.isRequired,
-      link: PropTypes.string.isRequired,
-    })
-  ),
-}
